implemented parallel_gates_grad_matfree and transpose_statevector; using integer typedef for iterating over statevector entries (32 bit could be insufficient)

+++ src/util.c
@@ -5,10 +5,10 @@
 ///
 /// \brief Uniform distance (infinity norm) between vectors 'x' and 'y'.
 ///
-double uniform_distance(int n, const numeric* x, const numeric* y)
+double uniform_distance(size_t n, const numeric* x, const numeric* y)
 {
 	double d = 0;
-	for (int i = 0; i < n; i++)
+	for (size_t i = 0; i < n; i++)
 	{
 		d = fmax(d, fabs(x[i] - y[i]));
 	}
@@ -21,10 +21,10 @@ double uniform_distance(int n, const numeric* x, const numeric* y)
 ///
 /// \brief Maximum element-wise relative distance of two vectors.
 ///
-double relative_distance(int n, const numeric* x, const numeric* y, double eps)
+double relative_distance(size_t n, const numeric* x, const numeric* y, double eps)
 {
 	double d = 0;
-	for (int i = 0; i < n; i++)
+	for (size_t i = 0; i < n; i++)
 	{
 		d = fmax(d, reldist(x[i], y[i], eps));
 	}
